fix: 토큰 만료 7일 전 갱신 기능 추가

+++ src/main/java/io/wanted/market/domain/token/TokenWriter.java
@@ -12,6 +12,10 @@ public class TokenWriter {
 
     @Transactional
     public Long write(TokenPair tokenPair) {
+        if (tokenPair.refreshToken() == null) {
+            return null;
+        }
+
         if (tokenRepository.existsByUserId(tokenPair.userId())) {
             tokenRepository.deleteByUserId(tokenPair.userId());
         }
